items: separate bead items and effects
